Determine whether a point is inside the 'funnel' of acceptable values

This function determines whether a given hydrographic value lies what the TEOS-10 literature calls a "funnel" of values that lead to acceptably accurate computation of specific volume. For more details, consult the TEOS-10 literature, perhaps starting with the materials referred to in the webpage cited in the ‘References’ section.

Usage

gsw_infunnel(SA, CT, p)

Arguments

SA

Absolute Salinity [ g/kg ]. The valid range for most `gsw` functions is 0 to 42 g/kg.

CT

Conservative Temperature [ degC ].

p

sea pressure [dbar], i.e. absolute pressure [dbar] minus 10.1325 dbar

Value

a logical value indicating whether the given point is inside the funnel of acceptable values.

References

https://www.teos-10.org/pubs/gsw/html/gsw_infunnel.html

See also

Examples

library(gsw)
gsw_infunnel(35, 10, 100) # TRUE
#> [1] TRUE
gsw_infunnel(45, 10, 100) # FALSE: too salty
#> [1] FALSE
gsw_infunnel(35, -4, 100) # FALSE: below freezing
#> [1] FALSE