Latent heat of evaporation

Latent heat of evaporation

gsw_latentheat_evap_CT(SA, CT)

Arguments

SA
Absolute Salinity [ g/kg ]

CT
Conservative Temperature [ degC ]

Value

latent heat of evaporation [ J/kg ]

Details

This R function uses a wrapper to a C function contained within the GSW-C system as updated 2021-mar-27 at https://github.com/TEOS-10/GSW-C with git commit f7bfebf44f686034636facb09852f1d5760c27f5. In turn, that C function uses data from the library/gsw_data_v3_0.mat file provided in the GSW-Matlab source code, version 3.06-11. Unfortunately, this version is no longer displayed on the teos-10.org website. Therefore, in the interests of making GSW-R be self-contained, a copy was downloaded from http://www.teos-10.org/software/gsw_matlab_v3_06_11.zip on 2021-jul-06, the .mat file was stored in the developer/create_data directory of https://github.com/TEOS-10/GSW-R, and then the dataset used in GSW-R was created based on that .mat file.

Please consult http://www.teos-10.org to learn more about these software systems, their authorships, and the science behind it all.

References

http://www.teos-10.org/pubs/gsw/html/gsw_latentheat_evap_CT.html

See also

Examples

SA <- c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
CT <- c(28.7856, 28.4329, 22.8103, 10.2600,  6.8863,  4.4036)
lh  <- gsw_latentheat_evap_CT(SA, CT)
stopifnot(all.equal(lh/1e6, c(2.429947107462561, 2.430774073049213, 2.444220372158452,
                              2.474127109232524, 2.482151446148560, 2.488052297193594)))