Potential density anomaly referenced to 3000 dbar

This uses the 75-term density equation, and returns potential density referenced to a pressure of 3000 dbar, minus 1000 kg/m^3.

gsw_sigma3(SA, CT)

Arguments

SA
Absolute Salinity [ g/kg ]

CT
Conservative Temperature [ degC ]

Value

potential density anomaly with reference pressure 3000 dbar [ kg/m^3 ]

Implementation Note

This R function uses a wrapper to a C function contained within the GSW-C system as updated 2021-03-27 at https://github.com/TEOS-10/GSW-C with git commit `f7bfebf44f686034636facb09852f1d5760c27f5`. In turn, that C function uses data from the library/gsw_data_v3_0.mat file provided in the GSW-Matlab source code, version 3.06-11. Unfortunately, this version is no longer displayed on the TEOS-10.org website. Therefore, in the interests of making GSW-R be self-contained, a copy was downloaded from http://www.teos-10.org/software/gsw_matlab_v3_06_11.zip on 2021-0706, the .mat file was stored in the developer/create_data directory of https://github.com/TEOS-10/GSW-R, and then the dataset used in GSW-R was created based on that .mat file. Please consult http://www.teos-10.org to learn more about the various TEOS-10 software systems.

References

http://www.teos-10.org/pubs/gsw/html/gsw_sigma3.html

See also

Examples

SA <- c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
CT <- c(28.8099, 28.4392, 22.7862, 10.2262,  6.8272,  4.3236)
sigma3 <- gsw_sigma3(SA,CT)
stopifnot(all.equal(sigma3, c(34.003747849903675, 34.267409891564057, 36.316415829697917,
                              39.732367693977039, 40.397934186745033, 40.881795690566832)))