Specific Volume of Ice

Specific Volume of Ice

gsw_specvol_ice(t, p)

Arguments

t
in-situ temperature (ITS-90) [ degC ]

p
sea pressure [dbar], i.e. absolute pressure [dbar] minus 10.1325 dbar

Value

Specific volume [ m^3/kg ]

Details

This R function uses a wrapper to a C function contained within the GSW-C system as updated 2021-mar-27 at https://github.com/TEOS-10/GSW-C with git commit f7bfebf44f686034636facb09852f1d5760c27f5. In turn, that C function uses data from the library/gsw_data_v3_0.mat file provided in the GSW-Matlab source code, version 3.06-11. Unfortunately, this version is no longer displayed on the teos-10.org website. Therefore, in the interests of making GSW-R be self-contained, a copy was downloaded from http://www.teos-10.org/software/gsw_matlab_v3_06_11.zip on 2021-jul-06, the .mat file was stored in the developer/create_data directory of https://github.com/TEOS-10/GSW-R, and then the dataset used in GSW-R was created based on that .mat file.

Please consult http://www.teos-10.org to learn more about these software systems, their authorships, and the science behind it all.

References

http://www.teos-10.org/pubs/gsw/html/gsw_specvol_ice.html

See also

Examples

t <- c(-10.7856, -13.4329, -12.8103, -12.2600,  -10.8863,  -8.4036)
p <- c(      10,       50,      125,      250,       600,     1000)
v <- gsw_specvol_ice(t, p)
stopifnot(all.equal(v, c(0.001088982980677, 0.001088489459509, 0.001088499019939,
                         0.001088433747301, 0.001088223220685, 0.001088135464776)))